feat:Amélioration tourelles, gain de crédits lors de la mort du monstre, BFS(Julien), Gestion interface amélioration tourelles

+++ TDJeu/src/application/modele/Monstre/Pig.java
@@ -2,7 +2,7 @@ package application.modele.Monstre;
 
 public class Pig extends Monstre {
 
-	public Pig(String nom, int x, int y) {
+	public Pig(String nom, double x, double y) {
 		super(nom, 12, 2, 50, x, y);
 	}	
 }
